+++ sources/Google.Solutions.IapDesktop.Extensions.Management/Views/InstanceControlCommands.cs
@@ -37,7 +37,6 @@ using System;
 using System.Diagnostics;
 using System.Threading.Tasks;
 using System.Windows.Forms;
-using Google.Solutions.IapDesktop.Core.ClientModel.Traits;
 
 namespace Google.Solutions.IapDesktop.Extensions.Management.Views
 {
